Unified pool snapshot interval (#533)
* add PoolSnapshot

* cleanup

* update config

* retreive decimals in pool deployed event

* update snapshot logic

* clean up

* refactor

---------
